Related work items: #3065, #3069, #3673, #3684, #3689, #3715

+++ Modules/Stations/AWG.Stations.handlers/CQRS/StationsCommandHandler.cs
@@ -35,7 +35,7 @@ namespace AWG.Stations.handlers.Command
       {
         station = mapper.Map<Model.Station>(request.Model);
 
-        station.Id = $"urn:ngsi-ld:Device:{Guid.NewGuid().ToString()}";
+        station.Id = $"urn:ngsi-ld:Device:{station.DataProvider}-{station.SerialNumber}";
         station.DateCreated = now;
         station.Category = new List<string>() { "sensor" };
         station.Owner = new List<string>();
@@ -53,6 +53,23 @@ namespace AWG.Stations.handlers.Command
 
       await mediator.Publish(new UpdateStationNotification() { Id = station.Id });
 
+      if (request.CBEnabled)
+      {
+        var entityId = await mediator.Send(new CreateOrUpdateCBEntity() { Id = station.Id });
+
+        if (String.IsNullOrEmpty(station.Source))
+        {
+          var subscriptionId = await mediator.Send(new SubscribeToCBEntity() { Id = entityId });
+
+          if (subscriptionId != null)
+          {
+            station.Source = subscriptionId;
+
+            await db.SaveChangesAsync();
+          }
+        }
+      }
+
       return mapper.Map<fiware.Device>(station);
     }
 
@@ -60,9 +77,17 @@ namespace AWG.Stations.handlers.Command
     {
       var station = await db.Stations.Where(f => f.Id == request.Id).FirstOrDefaultAsync();
 
+      if (station == null)
+        return Unit.Value;
+
       db.Stations.Remove(station);
 
       await db.SaveChangesAsync();
+
+      if (request.CBEnabled)
+      {
+        await mediator.Send(new UnsubscribeFromCBEntity() { SubscriptionId = station.Source });
+      }
 
       return Unit.Value;
     }
